requirements.txt
flask==1.1.2
flask-inputs==0.3.0
flask-jsontools==0.1.7
sqlalchemy==1.3.20
jsonschema==3.0.1
pytest==4.6.2
psycopg2==2.8.4
stripe==2.55.0
pyjwt==1.7.1
